+++ lib/src/ast/sass/statement/callable_declaration.dart
@@ -7,6 +7,7 @@ import 'package:source_span/source_span.dart';
 import '../argument_declaration.dart';
 import '../statement.dart';
 import 'parent.dart';
+import 'silent_comment.dart';
 
 /// An abstract class for callables (functions or mixins) that are declared in
 /// user code.
@@ -16,12 +17,17 @@ abstract class CallableDeclaration extends ParentStatement {
   /// This may be `null` for callables without names.
   final String name;
 
+  /// The comment immediately preceding this declaration.
+  final SilentComment comment;
+
   /// The declared arguments this callable accepts.
   final ArgumentDeclaration arguments;
 
   final FileSpan span;
 
   CallableDeclaration(
-      this.name, this.arguments, Iterable<Statement> children, this.span)
-      : super(List.unmodifiable(children));
+      this.name, this.arguments, Iterable<Statement> children, this.span,
+      {SilentComment comment})
+      : comment = comment,
+        super(List.unmodifiable(children));
 }