+++ lib/src/async_import_cache.dart
@@ -10,6 +10,7 @@ import 'package:tuple/tuple.dart';
 
 import 'ast/sass.dart';
 import 'importer.dart';
+import 'importer/result.dart';
 import 'io.dart';
 import 'logger.dart';
 import 'sync_package_resolver.dart';
@@ -33,6 +34,9 @@ class AsyncImportCache {
 
   /// The parsed stylesheets for each canonicalized import URL.
   final Map<Uri, Stylesheet> _importCache;
+
+  /// The import results for each canonicalized import URL.
+  final Map<Uri, ImporterResult> _resultsCache;
 
   /// Creates an import cache that resolves imports using [importers].
   ///
@@ -58,14 +62,16 @@ class AsyncImportCache {
       : _importers = _toImporters(importers, loadPaths, packageResolver),
         _logger = logger ?? const Logger.stderr(),
         _canonicalizeCache = {},
-        _importCache = {};
+        _importCache = {},
+        _resultsCache = {};
 
   /// Creates an import cache without any globally-avaiable importers.
   AsyncImportCache.none({Logger logger})
       : _importers = const [],
         _logger = logger ?? const Logger.stderr(),
         _canonicalizeCache = {},
-        _importCache = {};
+        _importCache = {},
+        _resultsCache = {};
 
   /// Converts the user's [importers], [loadPaths], and [packageResolver]
   /// options into a single list of importers.
@@ -96,7 +102,8 @@ class AsyncImportCache {
       : _importers = const [],
         _logger = const Logger.stderr(),
         _canonicalizeCache = const {},
-        _importCache = const {};
+        _importCache = const {},
+        _resultsCache = const {};
 
   /// Canonicalizes [url] according to one of this cache's importers.
   ///
@@ -177,6 +184,8 @@ Relative canonical URLs are deprecated and will eventually be disallowed.
     return await putIfAbsentAsync(_importCache, canonicalUrl, () async {
       var result = await importer.load(canonicalUrl);
       if (result == null) return null;
+
+      _resultsCache[canonicalUrl] = result;
       return Stylesheet.parse(result.contents, result.syntax,
           // For backwards-compatibility, relative canonical URLs are resolved
           // relative to [originalUrl].
@@ -189,8 +198,7 @@ Relative canonical URLs are deprecated and will eventually be disallowed.
 
   /// Return a human-friendly URL for [canonicalUrl] to use in a stack trace.
   ///
-  /// Throws a [StateError] if the stylesheet for [canonicalUrl] hasn't been
-  /// loaded by this cache.
+  /// Returns [canonicalUrl] as-is if it hasn't been loaded by this cache.
   Uri humanize(Uri canonicalUrl) {
     // Display the URL with the shortest path length.
     var url = minBy(
@@ -206,6 +214,12 @@ Relative canonical URLs are deprecated and will eventually be disallowed.
     return url.resolve(p.url.basename(canonicalUrl.path));
   }
 
+  /// Returns the URL to use in the source map to refer to [canonicalUrl].
+  ///
+  /// Returns [canonicalUrl] as-is if it hasn't been loaded by this cache.
+  Uri sourceMapUrl(Uri canonicalUrl) =>
+      _resultsCache[canonicalUrl]?.sourceMapUrl ?? canonicalUrl;
+
   /// Clears the cached canonical version of the given [url].
   ///
   /// Has no effect if the canonical version of [url] has not been cached.
@@ -218,6 +232,7 @@ Relative canonical URLs are deprecated and will eventually be disallowed.
   ///
   /// Has no effect if the imported file at [canonicalUrl] has not been cached.
   void clearImport(Uri canonicalUrl) {
+    _resultsCache.remove(canonicalUrl);
     _importCache.remove(canonicalUrl);
   }
 }