+++ test/synchronize_test.dart
@@ -10,14 +10,11 @@ import 'dart:io';
 import 'package:crypto/crypto.dart';
 import 'package:test/test.dart';
 
+import '../tool/grind/synchronize.dart' as synchronize;
+
 void main() {
   test("synchronized files are up-to-date", () {
-    ({
-      'lib/src/visitor/async_evaluate.dart': 'lib/src/visitor/evaluate.dart',
-      'lib/src/async_environment.dart': 'lib/src/environment.dart',
-      'lib/src/async_import_cache.dart': 'lib/src/import_cache.dart'
-    })
-        .forEach((sourcePath, targetPath) {
+    synchronize.sources.forEach((sourcePath, targetPath) {
       var source = File(sourcePath).readAsStringSync();
       var target = File(targetPath).readAsStringSync();